+++ lib/smart_core/types/primitive/invariant_control/single.rb
@@ -2,7 +2,7 @@
 
 # @api private
 # @since 0.2.0
-# @version 0.3.0
+# @version 0.8.0
 class SmartCore::Types::Primitive::InvariantControl::Single
   require_relative 'single/result'
 
@@ -45,6 +45,16 @@ class SmartCore::Types::Primitive::InvariantControl::Single
   def check(value, runtime_attributes)
     validation_result = !!invariant_checker.call(value, runtime_attributes)
     Result.new(self, value, validation_result)
+  end
+
+  # @param value [Any]
+  # @param runtime_attributes [Array<Any>]
+  # @return [Boolean]
+  #
+  # @api private
+  # @since 0.8.0
+  def simply_check(value, runtime_attributes)
+    !!invariant_checker.call(value, runtime_attributes)
   end
 
   private